Inline ImageComponent into ImageWidget
